requirements.txt
fastapi 
sqlalchemy 
psycopg2-binary 
pytest 
alembic